KVM: x86: Add KVM_X86_SNP_VM vm_type
commit 98ab5eba182296a3db55288d351ff61e271b2343 upstream.

In some cases, such as detecting whether a page fault should be handled
as a private fault or not, KVM will need to handle things differently
versus the existing KVM_X86_PROTECTED_VM type.

Add a new KVM_X86_SNP_VM to allow for this, along with a helper to query
the vm_type.

Signed-off-by: Michael Roth <michael.roth@amd.com>